同步

+++ eladmin-system/src/main/java/me/zhengjie/modules/system/rest/UserController.java
@@ -6,13 +6,15 @@ import me.zhengjie.domain.Picture;
 import me.zhengjie.domain.VerificationCode;
 import me.zhengjie.modules.system.domain.User;
 import me.zhengjie.exception.BadRequestException;
+import me.zhengjie.modules.system.domain.vo.UserPassVo;
 import me.zhengjie.modules.system.service.DeptService;
+import me.zhengjie.modules.system.service.RoleService;
+import me.zhengjie.modules.system.service.dto.RoleSmallDTO;
+import me.zhengjie.modules.system.service.dto.UserQueryCriteria;
 import me.zhengjie.service.PictureService;
 import me.zhengjie.service.VerificationCodeService;
 import me.zhengjie.utils.*;
 import me.zhengjie.modules.system.service.UserService;
-import me.zhengjie.modules.system.service.dto.UserDTO;
-import me.zhengjie.modules.system.service.query.UserQueryService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
@@ -25,9 +27,10 @@ import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 import java.util.*;
+import java.util.stream.Collectors;
 
 /**
- * @author jie
+ * @author Zheng Jie
  * @date 2018-11-23
  */
 @RestController
@@ -36,9 +39,6 @@ public class UserController {
 
     @Autowired
     private UserService userService;
-
-    @Autowired
-    private UserQueryService userQueryService;
 
     @Autowired
     private PictureService pictureService;
@@ -50,21 +50,21 @@ public class UserController {
     private DeptService deptService;
 
     @Autowired
+    private RoleService roleService;
+
+    @Autowired
     private VerificationCodeService verificationCodeService;
-
-
-    private static final String ENTITY_NAME = "user";
 
     @Log("查询用户")
     @GetMapping(value = "/users")
     @PreAuthorize("hasAnyRole('ADMIN','USER_ALL','USER_SELECT')")
-    public ResponseEntity getUsers(UserDTO userDTO, Pageable pageable){
+    public ResponseEntity getUsers(UserQueryCriteria criteria, Pageable pageable){
         Set<Long> deptSet = new HashSet<>();
         Set<Long> result = new HashSet<>();
 
-        if (!ObjectUtils.isEmpty(userDTO.getDeptId())) {
-            deptSet.add(userDTO.getDeptId());
-            deptSet.addAll(dataScope.getDeptChildren(deptService.findByPid(userDTO.getDeptId())));
+        if (!ObjectUtils.isEmpty(criteria.getDeptId())) {
+            deptSet.add(criteria.getDeptId());
+            deptSet.addAll(dataScope.getDeptChildren(deptService.findByPid(criteria.getDeptId())));
         }
 
         // 数据权限
@@ -78,14 +78,16 @@ public class UserController {
             result.retainAll(deptIds);
 
             // 若无交集，则代表无数据权限
+            criteria.setDeptIds(result);
             if(result.size() == 0){
                 return new ResponseEntity(PageUtil.toPage(null,0),HttpStatus.OK);
-            } else return new ResponseEntity(userQueryService.queryAll(userDTO,result,pageable),HttpStatus.OK);
+            } else return new ResponseEntity(userService.queryAll(criteria,pageable),HttpStatus.OK);
         // 否则取并集
         } else {
             result.addAll(deptSet);
             result.addAll(deptIds);
-            return new ResponseEntity(userQueryService.queryAll(userDTO,result,pageable),HttpStatus.OK);
+            criteria.setDeptIds(result);
+            return new ResponseEntity(userService.queryAll(criteria,pageable),HttpStatus.OK);
         }
     }
 
@@ -93,9 +95,7 @@ public class UserController {
     @PostMapping(value = "/users")
     @PreAuthorize("hasAnyRole('ADMIN','USER_ALL','USER_CREATE')")
     public ResponseEntity create(@Validated @RequestBody User resources){
-        if (resources.getId() != null) {
-            throw new BadRequestException("A new "+ ENTITY_NAME +" cannot already have an ID");
-        }
+        checkLevel(resources);
         return new ResponseEntity(userService.create(resources),HttpStatus.CREATED);
     }
 
@@ -103,6 +103,7 @@ public class UserController {
     @PutMapping(value = "/users")
     @PreAuthorize("hasAnyRole('ADMIN','USER_ALL','USER_EDIT')")
     public ResponseEntity update(@Validated(User.Update.class) @RequestBody User resources){
+        checkLevel(resources);
         userService.update(resources);
         return new ResponseEntity(HttpStatus.NO_CONTENT);
     }
@@ -111,38 +112,31 @@ public class UserController {
     @DeleteMapping(value = "/users/{id}")
     @PreAuthorize("hasAnyRole('ADMIN','USER_ALL','USER_DELETE')")
     public ResponseEntity delete(@PathVariable Long id){
+        Integer currentLevel =  Collections.min(roleService.findByUsers_Id(SecurityUtils.getUserId()).stream().map(RoleSmallDTO::getLevel).collect(Collectors.toList()));
+        Integer optLevel =  Collections.min(roleService.findByUsers_Id(id).stream().map(RoleSmallDTO::getLevel).collect(Collectors.toList()));
+
+        if (currentLevel > optLevel) {
+            throw new BadRequestException("角色权限不足");
+        }
         userService.delete(id);
         return new ResponseEntity(HttpStatus.OK);
     }
 
     /**
-     * 验证密码
-     * @param pass
-     * @return
-     */
-    @GetMapping(value = "/users/validPass/{pass}")
-    public ResponseEntity validPass(@PathVariable String pass){
-        UserDetails userDetails = SecurityContextHolder.getUserDetails();
-        Map map = new HashMap();
-        map.put("status",200);
-        if(!userDetails.getPassword().equals(EncryptUtils.encryptPassword(pass))){
-           map.put("status",400);
-        }
-        return new ResponseEntity(map,HttpStatus.OK);
-    }
-
-    /**
      * 修改密码
-     * @param pass
+     * @param user
      * @return
      */
-    @GetMapping(value = "/users/updatePass/{pass}")
-    public ResponseEntity updatePass(@PathVariable String pass){
-        UserDetails userDetails = SecurityContextHolder.getUserDetails();
-        if(userDetails.getPassword().equals(EncryptUtils.encryptPassword(pass))){
+    @PostMapping(value = "/users/updatePass")
+    public ResponseEntity updatePass(@RequestBody UserPassVo user){
+        UserDetails userDetails = SecurityUtils.getUserDetails();
+        if(!userDetails.getPassword().equals(EncryptUtils.encryptPassword(user.getOldPass()))){
+            throw new BadRequestException("修改失败，旧密码错误");
+        }
+        if(userDetails.getPassword().equals(EncryptUtils.encryptPassword(user.getNewPass()))){
             throw new BadRequestException("新密码不能与旧密码相同");
         }
-        userService.updatePass(userDetails.getUsername(),EncryptUtils.encryptPassword(pass));
+        userService.updatePass(userDetails.getUsername(),EncryptUtils.encryptPassword(user.getNewPass()));
         return new ResponseEntity(HttpStatus.OK);
     }
 
@@ -153,9 +147,8 @@ public class UserController {
      */
     @PostMapping(value = "/users/updateAvatar")
     public ResponseEntity updateAvatar(@RequestParam MultipartFile file){
-        UserDetails userDetails = SecurityContextHolder.getUserDetails();
-        Picture picture = pictureService.upload(file,userDetails.getUsername());
-        userService.updateAvatar(userDetails.getUsername(),picture.getUrl());
+        Picture picture = pictureService.upload(file, SecurityUtils.getUsername());
+        userService.updateAvatar(SecurityUtils.getUsername(),picture.getUrl());
         return new ResponseEntity(HttpStatus.OK);
     }
 
@@ -168,7 +161,7 @@ public class UserController {
     @Log("修改邮箱")
     @PostMapping(value = "/users/updateEmail/{code}")
     public ResponseEntity updateEmail(@PathVariable String code,@RequestBody User user){
-        UserDetails userDetails = SecurityContextHolder.getUserDetails();
+        UserDetails userDetails = SecurityUtils.getUserDetails();
         if(!userDetails.getPassword().equals(EncryptUtils.encryptPassword(user.getPassword()))){
             throw new BadRequestException("密码错误");
         }
@@ -176,5 +169,17 @@ public class UserController {
         verificationCodeService.validated(verificationCode);
         userService.updateEmail(userDetails.getUsername(),user.getEmail());
         return new ResponseEntity(HttpStatus.OK);
+    }
+
+    /**
+     * 如果当前用户的角色级别低于创建用户的角色级别，则抛出权限不足的错误
+     * @param resources
+     */
+    private void checkLevel(User resources) {
+        Integer currentLevel =  Collections.min(roleService.findByUsers_Id(SecurityUtils.getUserId()).stream().map(RoleSmallDTO::getLevel).collect(Collectors.toList()));
+        Integer optLevel = roleService.findByRoles(resources.getRoles());
+        if (currentLevel > optLevel) {
+            throw new BadRequestException("角色权限不足");
+        }
     }
 }
